infra: add kube-state-metrics and kubernetes-monitoring rules/alerts
tl;dr This adds alerts and more graphs.
- Alerts: https://infra.moov.io/prometheus/alerts
- Graphs: https://infra.moov.io/grafana/

kube-state-metrics is a kubernetes service which is designed to pick
up cluster stats. It's pretty common to see deployed.

Link: https://github.com/kubernetes/kube-state-metrics

kubernetes-monitoring is a Github org with some k8s and Prometheus
folks who are working on a good set of alerts and dashboards for k8s installs.

Link: https://github.com/kubernetes-monitoring/kubernetes-mixin