lb: split traefik into alpha/beta
We're just splitting the PVC and Deployment really, and it's just the
name of each having a suffix added. This means we'll need to update
both in the future, but I hope the filenames are explicit to be a reminder.

Also, set a pod anti-affinity of 'app: traefik' pods. That means we
won't be scheduled on a node that already has a pod with that label.

Fixes: https://github.com/moov-io/infra/issues/20